PAYARA-3075 Remove unused package and class
