+++ appserver/payara-appserver-modules/microprofile/fault-tolerance/src/main/java/fish/payara/microprofile/faulttolerance/FaultToleranceService.java
@@ -39,7 +39,6 @@
  */
 package fish.payara.microprofile.faulttolerance;
 
-import fish.payara.microprofile.faulttolerance.objects.FaultToleranceObject;
 import fish.payara.microprofile.faulttolerance.state.CircuitBreakerState;
 import fish.payara.notification.requesttracing.RequestTraceSpan;
 import fish.payara.nucleus.requesttracing.RequestTracingService;
